requirements.txt
Pillow
cs50